refactor config matching utils to reduce duplication
