+++ graphein/protein/config.py
@@ -11,45 +11,12 @@ from pathlib import Path
 from typing import Any, Callable, List, Optional, Union
 
 from deepdiff import DeepDiff
-from deepdiff.operator import BaseOperator
 from pydantic import BaseModel
 from typing_extensions import Literal
 
 from graphein.protein.edges.distance import add_peptide_bonds
 from graphein.protein.features.nodes.amino_acid import meiler_embedding
-
-
-def partial_functions_equal(func1: partial, func2: partial) -> bool:
-    """
-    Determine whether two partial functions are equal.
-
-    :param func1: Partial function to check
-    :type func1: partial
-    :param func2: Partial function to check
-    :type func2: partial
-    :return: Whether the two functions are equal
-    :rtype: bool
-    """
-    if not (isinstance(func1, partial) and isinstance(func2, partial)):
-        return False
-    return all(
-        getattr(func1, attr) == getattr(func2, attr)
-        for attr in ["func", "args", "keywords"]
-    )
-
-
-class PartialMatchOperator(BaseOperator):
-    """Custom operator for deepdiff comparison. This operator compares whether the two partials are equal."""
-
-    def give_up_diffing(self, level, diff_instance):
-        return partial_functions_equal(level.t1, level.t2)
-
-
-class PathMatchOperator(BaseOperator):
-    """Custom operator for deepdiff comparison. This operator compares whether the two pathlib Paths are equal."""
-
-    def give_up_diffing(self, level, diff_instance):
-        return level.t1 == level.t2
+from graphein.utils.config import PartialMatchOperator, PathMatchOperator
 
 
 class DSSPConfig(BaseModel):
